Gemini-Bot/requirements.txt
python-dotenv
flask
google-generativeai
fastapi
uvicorn